refactor(tests): extract mi_determinism fixture builder into helpers
Move build_paired_template, write_bam, and build_mixed_orientation_bam
from test_group_determinism.rs into a new shared helpers module
(tests/integration/helpers/mi_determinism_fixture.rs) so that both the
existing group/dedup determinism tests and the upcoming runall
determinism test can consume the same 40-position × 256-UMI ×
2-orientation fixture without drift.

test_group_determinism.rs is updated to import from the new module;
its private read_qname_mi and assert_mi_deterministic helpers are
unchanged.
